Break the admin/account circular import
graphql/account.ts imported getAdminAccountStats from
graphql/admin.ts, while admin.ts imported Account from
account.ts and referenced it at module top level when building
AdminAccountEdge.  In ESM that ordering is fragile: depending on
which module is loaded first, the reference to Account inside
admin.ts can read undefined and the schema builder throws at
startup.

Move the moderator-only Account.postCount and
Account.lastPostPublished field definitions out of the
Account.drizzleNode call in account.ts and re-attach them from
admin.ts via builder.drizzleObjectField, the same pattern
Account.invitationLinks and Account.invitees already use.  The
import graph is now strictly admin.ts → account.ts and the
loader/field plumbing stays cohesive in one file.

https://github.com/hackers-pub/hackerspub/pull/271#discussion_r3172904208
